Reverted http api changes
